refactor(backend): reorganize handler imports and update paths for consistency
