+++ backend/src/build-system/handlers/file-structure/index.ts
