+++ backend/src/build-system/handlers/file-structure/prompt.ts
