+++ backend/src/build-system/handlers/product-manager/product-requirements-document/prd.ts
@@ -1,6 +1,6 @@
 import { BuildHandler, BuildResult } from 'src/build-system/types';
 import { BuilderContext } from 'src/build-system/context';
-import { prompts } from './prompt/prompt';
+import { prompts } from './prompt';
 import { ModelProvider } from 'src/common/model-provider';
 import { Logger } from '@nestjs/common';
